Shift old auth policies UI into storage since its only being used there
