+++ apps/studio/components/interfaces/Database/Policies/PolicyTableRow/index.tsx
@@ -92,9 +92,13 @@ const PolicyTableRowComponent = ({
 
   const admonitionMessage = useMemo(() => getTableAdmonitionMessage(status), [status])
 
+  console.log({ status, exposedSchemas })
+
   return (
     <Card className={cn(isPubliclyReadable && 'border-warning-500')}>
-      <CardHeader className={cn('py-3 px-4', status !== 'secured' && 'border-b-0')}>
+      <CardHeader
+        className={cn('py-3 px-4', status !== 'secured' && status !== 'unknown' && 'border-b-0')}
+      >
         <PolicyTableRowHeader
           table={table}
           isLocked={isLocked}
