+++ apps/studio/components/interfaces/Storage/StoragePolicies/PolicyEditorModal/PolicyEditorModal.constants.ts
@@ -1,0 +1,6 @@
+export const POLICY_MODAL_VIEWS = {
+  SELECTION: 'SELECTION',
+  TEMPLATES: 'TEMPLATES',
+  EDITOR: 'EDITOR',
+  REVIEW: 'REVIEW',
+}
