+++ apps/studio/components/interfaces/Storage/StoragePolicies/StoragePolicies.tsx
@@ -16,14 +16,14 @@ import {
 import { GenericSkeletonLoader } from 'ui-patterns/ShimmeringLoader'
 
 import { formatPoliciesForStorage, UNGROUPED_POLICY_SYMBOL } from '../Storage.utils'
-import { StoragePoliciesBucketRow } from './StoragePoliciesBucketRow'
-import { BucketsPolicies, type SelectBucketPolicyForAction } from './StoragePoliciesBucketsSection'
-import { StoragePoliciesEditPolicyModal } from './StoragePoliciesEditPolicyModal'
+import { PolicyEditorModal } from './PolicyEditorModal'
 import type {
   PostgresPolicyCreatePayload,
   PostgresPolicyUpdatePayload,
-} from '@/components/interfaces/Database/Policies/Policies.types'
-import { PolicyEditorModal } from '@/components/interfaces/Database/Policies/PolicyEditorModal'
+} from './StoragePolicies.types'
+import { StoragePoliciesBucketRow } from './StoragePoliciesBucketRow'
+import { BucketsPolicies, type SelectBucketPolicyForAction } from './StoragePoliciesBucketsSection'
+import { StoragePoliciesEditPolicyModal } from './StoragePoliciesEditPolicyModal'
 import type { Policy } from '@/components/interfaces/Database/Policies/PolicyTableRow/PolicyTableRow.utils'
 import { useDatabasePoliciesQuery } from '@/data/database-policies/database-policies-query'
 import { useDatabasePolicyCreateMutation } from '@/data/database-policies/database-policy-create-mutation'
