+++ apps/studio/components/interfaces/Storage/StoragePolicies/StoragePolicies.types.ts
@@ -58,3 +58,8 @@ export interface DraftPostgresPolicyUpdatePayload extends Omit<
   definition?: DisplayableSqlFragment
   check?: DisplayableSqlFragment
 }
+
+export interface PolicyForReview {
+  description?: string
+  statement?: string
+}
